requirements.txt
altgraph==0.17.4
ane_transformers==0.1.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
coremltools==7.2
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
macholib==1.16.3
MarkupSafe==2.1.5
modulegraph==0.19.6
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
openai==1.35.3
openai-whisper==20231117
packaging==24.1
protobuf==3.20.1
py2app==0.28.8
pyaml==24.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sniffio==1.3.1
srt==3.5.3
sympy==1.12.1
tiktoken==0.7.0
tkinterdnd2-universal==1.7.3
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
urllib3==2.2.2